+++ src/RcppExports.cpp
@@ -65,6 +65,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// is_date_format_
+LogicalVector is_date_format_(CharacterVector formats);
+RcppExport SEXP _tidyxl_is_date_format_(SEXP formatsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< CharacterVector >::type formats(formatsSEXP);
+    rcpp_result_gen = Rcpp::wrap(is_date_format_(formats));
+    return rcpp_result_gen;
+END_RCPP
+}
 // xlex_
 List xlex_(CharacterVector x);
 RcppExport SEXP _tidyxl_xlex_(SEXP xSEXP) {
@@ -83,6 +94,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_tidyxl_xlsx_sheet_files_", (DL_FUNC) &_tidyxl_xlsx_sheet_files_, 1},
     {"_tidyxl_xlsx_validation_", (DL_FUNC) &_tidyxl_xlsx_validation_, 3},
     {"_tidyxl_xlsx_names_", (DL_FUNC) &_tidyxl_xlsx_names_, 1},
+    {"_tidyxl_is_date_format_", (DL_FUNC) &_tidyxl_is_date_format_, 1},
     {"_tidyxl_xlex_", (DL_FUNC) &_tidyxl_xlex_, 1},
     {NULL, NULL, 0}
 };