+++ src/xlsxstyles.cpp
@@ -4,6 +4,7 @@
 #include "xlsxstyles.h"
 #include "xf.h"
 #include "color.h"
+#include "date.h"
 
 using namespace Rcpp;
 
@@ -268,41 +269,6 @@ void xlsxstyles::cacheNumFmts(rapidxml::xml_node<>* styleSheet) {
   numFmts_ = formatCodes;
   isDate_ = isDate;
 }
-
-// From @reviewher https://github.com/tidyverse/readxl/issues/388
-#define CASEI(c) case c: case c | 0x20
-#define CMPLC(j,n) if(x[i+j] | 0x20 == n)
-bool xlsxstyles::isDateFormat(std::string x) {
-  char escaped = 0;
-  char bracket = 0;
-  for (size_t i = 0; i < x.size(); ++i) switch (x[i]) {
-    CASEI('D'):
-    CASEI('E'):
-    CASEI('H'):
-    CASEI('M'):
-    CASEI('S'):
-    CASEI('Y'):
-      if(!escaped && !bracket) return true;
-      break;
-    case '"':
-      escaped = 1 - escaped; break;
-    case '\\': ++i; break;
-    case '[': if(!escaped) bracket = 1; break;
-    case ']': if(!escaped) bracket = 0; break;
-    CASEI('G'):
-      if(i + 6 < x.size())
-      CMPLC(1,'e')
-      CMPLC(2,'n')
-      CMPLC(3,'e')
-      CMPLC(4,'r')
-      CMPLC(5,'a')
-      CMPLC(6,'l')
-        return false;
-  }
-  return false;
-}
-#undef CMPLC
-#undef CASEI
 
 void xlsxstyles::cacheFonts(rapidxml::xml_node<>* styleSheet) {
   rapidxml::xml_node<>* fonts = styleSheet->first_node("fonts");